requirements.txt
aiogram==2.19
asyncio==3.4.3
aiohttp==3.8.1
python-dotenv==0.19.2
asyncpg==0.25.0
SQLAlchemy==1.3.24
gino==1.0.1